refactor:common.js 파일 생성해서 오류 해결
